+++ app/vmui/packages/vmui/src/components/Layout/Header/Header.tsx
@@ -59,11 +59,6 @@ const Header: FC = () => {
 
   const [activeMenu, setActiveMenu] = useState(pathname);
 
-  const handleChangeTab = (value: string) => {
-    setActiveMenu(value);
-    navigate(value);
-  };
-
   const headerSetup = useMemo(() => {
     return ((routerOptions[pathname] || {}) as RouterOptions).header || {};
   }, [pathname]);
@@ -91,7 +86,7 @@ const Header: FC = () => {
   >
     {!appModeEnable && (
       <div
-        className="vm-header__logo"
+        className="vm-header-logo"
         onClick={onClickLogo}
         style={{ color }}
       >
@@ -100,10 +95,10 @@ const Header: FC = () => {
     )}
     <div className="vm-header-nav">
       <Tabs
+        isNavLink
         activeItem={activeMenu}
         items={routes.filter(r => !r.hide)}
         color={color}
-        onChange={handleChangeTab}
       />
     </div>
     <div className="vm-header__settings">